requirements.txt
gazu
python-decouple
pika
